Make oem be a superset of vmx
A vmx image is the same disk as an oem just without the dracut
repart/resize feature. This difference is better handled with
an oemconfig parameter <oem-resize> which allows to switch resize
on or off. The extra image type vmx will be dropped and an XSLT
stylesheet automatically transforms a vmx description to be a
oem image with the resize feature switched off.

This Fixes #1425

!!! WIP don't merge !!!

- missing doc adaptions (oem = original equipment manufacturer)
  (Component disk image used in external workloads)
  check docs to make clear why disk images takes the name oem